backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
stripe==8.3.0
google-generativeai==0.3.2
anthropic==0.53.0
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
structlog==23.2.0
rich==13.7.0
websockets==15.0.1
PyYAML==6.0.2 